add new feature error handler customize.

+++ src/main/kotlin/com/jackcomeback/jinhan/board/repo/UserRepository.kt
@@ -4,7 +4,7 @@ import com.jackcomeback.jinhan.board.entity.User
 import org.springframework.data.repository.CrudRepository
 import java.util.*
 
-interface UserRepository : CrudRepository<User,Int> {
+interface UserRepository : CrudRepository<User,Long> {
     fun save(user : User) : User?
     fun findByName(name: String) : User?
 }
